Added handling of 64 bits fundamental types.
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@4358 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ common/utils/itti_analyzer/libparser/fundamental_type.h
@@ -3,7 +3,7 @@
 #ifndef FUNDAMENTAL_TYPE_H_
 #define FUNDAMENTAL_TYPE_H_
 
-uint32_t fundamental_read_from_buffer(struct types_s *type, buffer_t *buffer, uint32_t offset, uint32_t parent_offset);
+uint64_t fundamental_read_from_buffer(struct types_s *type, buffer_t *buffer, uint32_t offset, uint32_t parent_offset);
 
 int fundamental_dissect_from_buffer(
     struct types_s *type, ui_set_signal_text_cb_t ui_set_signal_text_cb, gpointer user_data,
